Organize information to create haplotypes using segregating sites matrices

This function organizes higher-level information for creating haplotypes from matrices of segregating sites output from coalescent simulations.

haps_ssites(obj = NULL, fn = NULL)

Arguments

obj

Object containing segregating sites information. This can be one of the following: (1) A single list with a seg_sites field inside. This field must contain a matrix for segregating sites for each chromosome. The matrix itself should contain the haplotype information, coded using 0s and 1s: 0s indicate the ancestral state and 1s indicate mutant. The matrix column names should indicate the positions of the polymorphisms on the chromosome. If positions are in the range (0,1), they're assumed to come from an infinite- sites model and are relative positions. If positions are integers in the range [0, chromosome length - 1] or [1, chromosome length], they're assumed to come from an finite-sites model and are absolute positions. Defaults to NULL.

fn

A single string specifying the name of the file containing the ms-style coalescent output with segregating site info. Defaults to NULL.

Value

A haps_ssites_info object containing information used in create_haplotypes to create variant haplotypes. This class is just a wrapper around a list of matrices of segregating site info, which you can view (but not change) using the object's mats() method.

Details

For what the seg_sites field should look like in a list, see output from the scrm or coala package. (These packages are not required to be installed when installing jackalope.) If using either of these packages, I encourage you to cite them. For citation information, see output from citation("scrm") or citation("coala").