An R6 class representing haploid variants from a reference genome.

This class should NEVER be created using variants$new. Only use create_variants. This class wraps a pointer to a C++ object, which is why there are no fields to manipulate directly. All manipulations are done through this class's methods.

variants

Format

An R6Class generator object

Value

An object of class variants.

Connections to ref_genome objects

Regarding the ref_genome object you use to create a variants object, you should note the following:

This point is the most important. Both the ref_genome and variants objects use the same underlying C++ object to store reference genome information. Thus, if you make any changes to the ref_genome object, those changes will also show up in the variants object. For example, if you make a variants object named V based on an existing ref_genome object named R, then you merge chromosomes in R, V will now have merged chromosomes. If you've already started adding mutations to V, then all the indexes used to store those mutations will be inaccurate. So when you do anything with V later, your R session will crash or have errors. The lesson here is that you shouldn't edit the reference genome after using it to create variants.

If a ref_genome object is used to create a variants object, deleting the ref_genome object won't cause issues with the variants object. However, the variants class doesn't provide methods to edit chromosomes, so only remove the ref_genome object when you're done editing the reference genome.

Methods

Viewing information:

ptr()

View the pointer to the variant information. This is used internally by jackalope and shouldn't be of much use to users.

n_chroms()

View the number of chromosomes.

n_vars()

View the number of variants.

sizes(var_ind)

View vector of chromosome sizes for a given variant.

chrom_names()

View vector of chromosome names.

var_names()

View vector of variant names.

chrom(var_ind, chrom_ind)

View a chromosome sequence string based on indices for the chromosome (chrom_ind) and variant (var_ind).

gc_prop(var_ind, chrom_ind, start, end)

View the GC proportion for a range within a variant chromosome.

nt_prop(nt, var_ind, chrom_ind, start, end)

View the proportion of a range within a variant chromosome that is of nucleotide nt.

Editing information:

set_names(new_names)

Set names for all variants. new_names is a character vector of what to change names to, and it must be the same length as the # variants.

add_vars(new_names)

Add new, named variant(s) to the object. These variants will have no mutations. If you want to add new variants with mutations, either re-run create_variants or use the dup_vars method to duplicate existing variants.

dup_vars(var_names, new_names = NULL)

Duplicate existing variant(s) based on their name(s). You can optionally specify the names of the duplicates (using new_names). Otherwise, their names are auto-generated.

rm_vars(var_names)

Remove one or more variants based on names in the var_names vector.

add_sub(var_ind, chrom_ind, pos, nt)

Manually add a substitution for a given variant (var_ind), chromosome (chrom_ind), and position (pos). The reference nucleotide will be changed to nt, which should be a single character.

add_ins(var_ind, chrom_ind, pos, nts)

Manually add an insertion for a given variant (var_ind), chromosome (chrom_ind), and position (pos). The nucleotide(s) nts will be inserted after the designated position.

add_del(var_ind, chrom_ind, pos, n_nts)

Manually add a deletion for a given variant (var_ind), chromosome (chrom_ind), and position (pos). The designated number of nucleotides to delete (n_nts) will be deleted starting at pos, unless pos is near the chromosome end and doesn't have n_nts nucleotides to remove; it simply stops at the chromosome end in this case.

See also